requirements.txt
google-generativeai
pypdf
cryptography